requirements.txt
gspread
google-auth
google-api-python-client
gspread-dataframe
python-dotenv
numpy
pandas